% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pounders}
\alias{pounders}
\title{Use Pounders to minimize a non-linear sum of squares problem}
\usage{
pounders(objFun, startValues, k, n)
}
\arguments{
\item{objFun}{is an R objective function that maps k parameters into n equations.}

\item{startValues}{is a vector with k elements}

\item{k}{is the number of parameters}

\item{n}{is the number of elements in the objective function}
}
\value{
a list with the objective function and the final parameter values
}
\description{
Use Pounders to minimize a non-linear sum of squares problem
}
\examples{
objfun = function(x) c((x[1] - 3), (x[2] + 1))
ret = pounders(objfun, c(1,2), 2, 2)
ret$x
}

