% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tao.R
\name{tao.optim}
\alias{tao.optim}
\title{R bindings for the TAO optimization library.}
\usage{
tao.optim(par, fn, gr = NULL, hs = NULL, method = c("nm", "pounders",
  "lmvm", "blmvm"), control = list(), n = 1)
}
\arguments{
\item{par}{Initial values for the parameters to be optimized over.}

\item{fn}{A function to be minimized (or maximized), with first argument
the vector of parameters over which minimization is to take place.
It should return a scalar result.}

\item{gr}{A function to return the gradient, if using a gradient-based
optimization method.}

\item{hs}{A function to return the hessian, if using an algorithm which
uses the hessian}

\item{method}{The method to be used. See 'Details'.}

\item{control}{A list of control parameters. See 'Details'.}

\item{n}{The number of elements of objfun.}
}
\value{
A list with final parameter values, the objective function, and
       information on why the optimizer stopped.
}
\description{
Here are some more details.
}
\details{
Further details.
}
\examples{
# Gradient-free method
objfun = function(x) c((x[1] - 3), (x[2] + 1))
ret = tao.optim(c(1, 2),
                objfun,
                method = "pounders",
                control = list(tao_pounders_delta="0.1"),
                n = 2)

# Gradient-based method
objfun = function(x) sum(c((x[1] - 3)^2, (x[2] + 1))^2)
grafun = function(x) c(2*(x[1] - 3), 2*(x[2] + 1))

ret = tao.optim(c(1, 2),
                objfun,
                gr = grafun,
                method = "lmvm")
}

