% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tao_cpp}
\alias{tao_cpp}
\title{Use TAO to minimize an objective function}
\usage{
tao_cpp(functions, start_values, method, options, n, lower_bounds, upper_bounds)
}
\arguments{
\item{functions}{is a list of Rcpp functions. The first is always the objective 
function. The second and third are optionally the Jacobian and the Hessian 
functions.}

\item{start_values}{is a vector containing the starting values of the parameters.}

\item{method}{is a string that determines the type of optimizer to be used.}

\item{options}{is a list containing option values for the optimizer}

\item{n}{is the number of elements in the objective function.}

\item{lower_bounds}{is a vector with lower bounds}

\item{upper_bounds}{is a vector with upper bounds}
}
\value{
a list with the objective function and the final parameter values
}
\description{
\code{tao_cpp} is an internal function of this package. It is recommended that
users call \code{\link{tao}} instead, which has a more convenient syntax and performs
thorough input checking.
}
\examples{
# use pounders
objfun = function(x) c((x[1] - 3), (x[2] + 1))
ret = tao_cpp(functions = list(objfun = objfun), 
              start_values = c(1, 2), 
              method = "pounders", 
              options = list(), 
              n = 2,
              lower_bounds = c(-2, -2),
              upper_bounds = c(5, 5))
ret$x
    
# use Nelder-Mead
objfun = function(x) sum(c((x[1] - 3)^2, (x[2] + 1))^2)
ret = tao_cpp(functions = list(objfun = objfun), 
                  start_values = c(1, 2), 
                  method = "nm", 
                  options = list(),
                  n = 1,
                  lower_bounds = c(-2, -2),
                  upper_bounds = c(5, 5))
ret$x
}

