% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tao}
\alias{tao}
\title{Use Pounders to minimize a non-linear sum of squares problem}
\usage{
tao(objFun, startValues, optimizer, k, n = 1L)
}
\arguments{
\item{objFun}{is an R objective function that maps k parameters into n equations.}

\item{startValues}{is a vector with k elements}

\item{optimizer}{is a string that determines the type of optimizer to be used. 
This needs to be one of \code{nm}, \code{pounders}, \code{lmvm}, \code{blmvm}.}

\item{k}{is the number of parameters}

\item{n}{is the number of elements in the objective function}
}
\value{
a list with the objective function and the final parameter values
}
\description{
Use Pounders to minimize a non-linear sum of squares problem
}
\examples{
# use pounders
objfun = function(x) c((x[1] - 3), (x[2] + 1))
ret = tao(objfun, c(1,2), "pounders", 2, 2)
ret$x
    
# use Nelder-Mead
objfun = function(x) sum(c((x[1] - 3), (x[2] + 1))^2)
ret = tao(objfun, c(1,2), "nm", 2)
ret$x
}

