% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tao}
\alias{tao}
\title{Use TAO to minimize an objective function}
\usage{
tao(functions, startValues, method, options, n = 1L)
}
\arguments{
\item{functions}{is a list of Rcpp functions. The first is always the objective
function. The second and third are optionally the Jacobian and the Hessian
functions.}

\item{startValues}{is a vector containing the starting values of the parameters.}

\item{method}{is a string that determines the type of optimizer to be used.}

\item{options}{is a list containing option values for the optimizer}

\item{n}{is the number of elements in the objective function.}
}
\value{
a list with the objective function and the final parameter values
}
\description{
Use TAO to minimize an objective function
}
\examples{
# use pounders
objfun = function(x) c((x[1] - 3), (x[2] + 1))
ret = tao(functions = list(objfun = objfun),
              startValues = c(1, 2),
              method = "pounders",
              options = list(),
              n = 2)
ret$x

# use Nelder-Mead
objfun = function(x) sum(c((x[1] - 3)^2, (x[2] + 1))^2)
ret = tao(functions = list(objfun = objfun),
                  startValues = c(1, 2),
                  method = "nm",
                  options = list())
ret$x
}

